requirements.txt
dash_html_components==1.0.3
requests==2.23.0
pandas==0.25.3
tqdm==4.42.1
scikit_image==0.16.2
dash_bootstrap_components==0.10.2
dash==1.12.0
numpy==1.17.3
dash_core_components==1.10.0
Pillow==7.2.0
scikit_learn==0.23.1
skimage==0.0
tensorflow==2.2.0
wikipedia==1.4.0
